requirements.txt
fastapi
uvicorn
requests
python-dotenv
pytest
jinja2
httpx
